requirements.txt
streamlit
PyPDF2
pdfplumber
tabula-py
pandas
